src/third_party/blink/web_tests/virtual/controls-refresh/calendar-picker/date-picker-hotkeys.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../fast/forms/resources/common.js"></script>
<script src="../../../fast/forms/resources/picker-common.js"></script>
<script src="../../../fast/forms/calendar-picker/resources/calendar-picker-common.js"></script>
</head>
<body>
<input type="date" id="date-0" value="2019-02-14">
<script>

promise_test(() => {
  let dateElement = document.getElementById('date-0');
  return openPickerWithPromise(dateElement)
  .then(() => {
    eventSender.keyDown('PageDown');
    assert_equals(dateElement.value, "2019-03-14", "Next month hotkey should have incremented month");

    eventSender.keyDown('PageUp');
    assert_equals(dateElement.value, "2019-02-14", "Previous month hotkey should have decremented month");

    eventSender.keyDown('Enter');
    assert_equals(internals.pagePopupWindow, null, "Enter key should have closed popup.");
  });
}, "Date picker: Previous/next month hotkey (PageDown/PageUp)");

</script>
</body>
</html>
